Added automatic image versions
Also, you don't have to type in the path of the image anymore as it will just look in the images folder. However, the current method might be difficult to scale as it looks for each valid image format separately when you type in just the name of the file.